Restructure package structure.
