+++ src/test/java/com/github/dakusui/cmd/ut/CmdTest.java
@@ -1,8 +1,9 @@
-package com.github.dakusui.streamablecmd.ut;
+package com.github.dakusui.cmd.ut;
 
-import com.github.dakusui.cmd.TestUtils;
-import com.github.dakusui.streamablecmd.Cmd;
-import com.github.dakusui.streamablecmd.exceptions.CommandExecutionException;
+import com.github.dakusui.cmd.utils.TestUtils;
+import com.github.dakusui.cmd.Cmd;
+import com.github.dakusui.cmd.Shell;
+import com.github.dakusui.cmd.exceptions.CommandExecutionException;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -41,11 +42,16 @@ public class CmdTest extends TestUtils.StdOutTestBase {
     };
   }
 
+  @Test
+  public void given$when$then() {
+    Cmd.run(Shell.local(), "echo hello && exit 1").forEach(System.out::println);
+  }
+
   @Test(expected = CommandExecutionException.class)
   public void main2() throws IOException {
     try {
       new Cmd.Builder()
-          .withShell(new Cmd.Shell.Builder.ForLocal().build())
+          .withShell(new Shell.Builder.ForLocal().build())
           .add("echo $(which echo) && echo \"hello\" && cat hello")
           .configure(defaultIo)
           .build()
@@ -63,9 +69,9 @@ public class CmdTest extends TestUtils.StdOutTestBase {
     try {
       new Cmd.Builder()
           .withShell(
-              new Cmd.Shell.Builder.ForSsh("localhost")
+              new Shell.Builder.ForSsh("localhost")
                   .userName(TestUtils.userName())
-                  .identity(TestUtils.userName())
+                  .identity(TestUtils.identity())
                   .build()
           )
           .add("echo")
@@ -85,7 +91,7 @@ public class CmdTest extends TestUtils.StdOutTestBase {
   public void main4() throws IOException {
     try {
       new Cmd.Builder()
-          .withShell(new Cmd.Shell.Builder.ForLocal().build())
+          .withShell(new Shell.Builder.ForLocal().build())
           .configure(createIo(Stream.of("hello", "world", "everyone")))
           .add("cat -n")
           .build()
@@ -102,7 +108,7 @@ public class CmdTest extends TestUtils.StdOutTestBase {
   public void main1() throws IOException {
     try {
       Cmd cmd = new Cmd.Builder()
-          .withShell(new Cmd.Shell.Builder.ForLocal().withProgram("sh").addOption("-c").build())
+          .withShell(new Shell.Builder.ForLocal().withProgram("sh").addOption("-c").build())
           .add(String.format("cat /dev/zero | head -c 100000 | %s 80", base64()))
           .configure(createIo(Stream.of("Hello", "world")))
           .build();
